fix: コントリビューション（草）の表示
